Don't use possible number pattern to determine number type
The phone number type determination should never use the possible
number pattern, since that pattern is generic and not necessarily
specific to the type in question.

Update `isNumberMatchingDesc` to always use the national number pattern,
not the possible number pattern.

Also found and fixed a bug in RegularExpressions matchesEntirely while
making this change.