Fix PartialFormatter for regions that are not the main country
The metadata number formats are only specified for the country
code's main country. Since PartialFormatter was using the metadata
of the region code, a region that is not the code's main country
would not have a format and PartialFormatter would leave the input
unchanged.

Fix by always using the main country for the region's country code.
This matches the behavior of libphonenumber.