Updated to Swift 3
This was done with the combination of the migrator, fix-mes and manual tweaks. The only issue that I ran into was that dispatch_once is no longer available. But looking at the place we were using it, it was an incorrect use of it to begin with. dispatch_once has undefined behavior when the token is not a global constant. Instead, I switched it to a basic guard statement based on state. This is not as thread safe necessarily, but it doesn't look like the rest of the operation takes great pains to be thread safe anyway.