Ensure convenience initialization works
We need to ensure that an instance of a PhoneNumberKit class is strongly referenced when creating a PartialFormatter, otherwise the weakly held properties will become nil, and formatting will fail. Allow for a default initialized parameter allows for all existing initialization sequences to work as they did before while allowing for convenience initialized PartialFormatters to work as well.