Xcode 10.2: update localization configuration for 0 files
opening this project in Xcode 10.2 results in issue navigator
diagnostics:
- a warning diagnostic that the project needs to be migrated
  w.r.t. the deprecated usage of "developmentRegion = English"
- another diagnostic w.r.t. the use of Base localization

however, there are 0 localized files in the project, so it is
actually possible to remove the localized "known region" (as
seen in the XML diff) to contain no regions.

Xcode 10.2.1 does suggest enabling the warning
"CLANG_ANALYZER_LOCALIZABILITY_NONLOCALIZED = YES"
in order to capture any cases of broken usage
(of which there are none currently).